Add mutually exclusive argument support
The mutually exclusive group can be added using
ArgumentParser#addMutuallyExclusiveGroup. The returned object
from this method implements MutuallyExclusiveGroup interface. It
has similar interface with ArgumentGroup, but it additionally has
required() method, which toggles the restriction that one of the
argument in this group must be specified or not.  Usually, more
than 1 argument are added to mutually exclusive group and only
one of them can be specified. The arguments are specially
formatted in usage and help messages.