Add additional help message for unrecognized args from file
If trailing empty lines or white spaces exist in argument file
accidentally, the error message is typically only "unrecognized
arguments" and it may be hard to notice that the problem is in a
file. To help the user in this situation, if the unrecognized
argument was coming from a file, add additional help message to
the exception message to alert user that the problem may exist in
argument file. This change also quotes unrecognized arguments in
exception message.