Adjust narg("*") handling for optinal arguments
The optional argument with nargs("*") and without default set is
not given in command-line, the result is now null instead of
empty list.

The optional argument with nargs("*") is given in command-line
with emtpy list, the result is now empty list. It will overwrite
existing value or appened to the existing list in case of
append() action.