Add ReflectArgumentType
ReflectArgumentType is an enhanced version of ConstructorArgumentType.
It first tries to convert given String to desired type using valueOf
static method of the type. If it failed, use the constructor of given
type just like a ConstructorArgumentType. Since enums have valueOf
static method, ReflectArgumentType works with enums as well.
The ReflectArgumentType effectively deprecates
ConstructorArgumentType and EnumArgumentType.