Introduce MetavarInference interface, and infer Boolean metavar
Previously, when Boolean.class is passed to Argument#type(), and no
metavar is provided, the help message was like "-f F".  This is
intentional, and expected, but it is better for user to show them F is
true or false.  This commit introduce MetavarInference interface.
When sub class of ArgumentType also implements MetavarInference, it
can infer metavar from its interface.  We made ReflectArgumentType
implement this interface, and if Boolean.class is given for its
constructor, then return useful Boolean value metavar from
inferMetavar() method.